+++ Crimson/src/Platform/Metal/MetalDeferredRenderer.h
@@ -15,7 +15,7 @@ namespace Crimson {
         
         static Ref<Shader> GetDeferredShader() { return m_DefferedPassShader; }
         
-        static uint32_t GetBuffers(uint32_t bufferInd);
+        static uintptr_t GetBuffers(uint32_t bufferInd);
 
     private:
         static void RenderEntities(Scene* scene);
